+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.96.0] - 2022-02-24
+### Fixed
+* Bug: Error in replacing property using config.csv when empty element exists #186
+* Bug: Replacing properties from json-config for managedPrivateEndpoints #184 #185
+
 ## [0.95.1] - 2021-12-07
 * Bug fixed: Loading ADF from code fails when excluding any object and ManagedVNET exist #171